+++ Bootstrap-GitLabRunner.ps1
@@ -43,7 +43,7 @@
 # Edit ONLY these values to match your environment.
 # ============================================================
 
-$Script:BootstrapVersion   = '2.4.5+pending'
+$Script:BootstrapVersion   = '2.4.6+pending'
 $Script:BootstrapEndpoint  = 'https://kayhut-minio.com:9000'
 $Script:BootstrapBucket    = 'gitlab-runner-golden'
 $Script:BootstrapAccessKey = 'YOUR_ACCESS_KEY_HERE'